Added skip text button
