+++ lib/features/on_boarding/ui/screens/on_boarding_screen.dart
@@ -1,4 +1,4 @@
-import 'package:bug_tracking/features/on_boarding/ui/widgets/custom_text_button.dart';
+import 'package:bug_tracking/features/on_boarding/ui/widgets/skip_text_button.dart';
 import 'package:flutter/material.dart';
 
 class OnBoardingScreen extends StatelessWidget {
@@ -8,15 +8,8 @@ class OnBoardingScreen extends StatelessWidget {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        actions: [
-          CustomTextButton(
-            onPressed: () {},
-            text: 'Skip',
-            textStyle: const TextStyle(
-              color: Colors.black,
-              fontSize: 18,
-            ),
-          ),
+        actions: const [
+          SkipTextButton(),
         ],
       ),
     );
